requirements.txt

Flask 
Authlib 
Flask-OAuthlib
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv